// apps/desktop/src/generated/query/BindingCommandResultDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BindingDto } from "./BindingDto";
import type { SnapshotMetadata } from "./SnapshotMetadata";

export type BindingCommandResultDto = { metadata: SnapshotMetadata, binding: BindingDto, };