// apps/desktop/src/generated/query/BindingDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BindingStatusDto } from "./BindingStatusDto";

export type BindingDto = { binding_id: string, source_resource_id: string, target_resource_id: string, kind: string, status: BindingStatusDto, created_at: string, updated_at: string, };