// apps/desktop/src/generated/query/ResourceDetailDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ChangeDto } from "./ChangeDto";
import type { ConnectorCoverageDto } from "./ConnectorCoverageDto";
import type { RelationDto } from "./RelationDto";
import type { ResourceDto } from "./ResourceDto";
import type { SnapshotMetadata } from "./SnapshotMetadata";

export type ResourceDetailDto = { metadata: SnapshotMetadata, resource: ResourceDto, attributes: unknown, relations: Array<RelationDto>, recent_changes: Array<ChangeDto>, connector_coverage: Array<ConnectorCoverageDto>, };